Guardar cambios

+++ src/controller/tools/BillPrintable.java
@@ -410,7 +410,7 @@ public class BillPrintable implements Printable {
         return bArray;
     }
 
-    private TextFieldTicket addElementTextField(String id, String titulo, boolean multilinea, short lines, int widthColumn, Pos align, boolean editable, String variable) {
+    private TextFieldTicket addElementTextField(String id, String titulo, boolean multilinea, short lines, short widthColumn, Pos align, boolean editable, String variable) {
         TextFieldTicket field = new TextFieldTicket(titulo, id);
         field.setMultilineas(multilinea);
         field.setLines(lines);
